int can overflow for large memory allocations
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@8381 6419b465-d902-0410-997c-be2266baf723

+++ c/extensions.h
@@ -34,7 +34,7 @@
 
 #define INLINE inline
 
-static INLINE void* gpaw_malloc(int n)
+static INLINE void* gpaw_malloc(size_t n)
 {
   void* p = malloc(n);
   assert(p != NULL);
